events pulled from db
