+++ app/routes/admin/events/data.server.ts
@@ -1,0 +1,9 @@
+import { db } from "~/services/db/db.server";
+
+const getEvents = async () => {
+  const events = await db.query.events.findMany({});
+
+  return { events };
+};
+
+export { getEvents };
