+++ app/services/db/db.server.ts
@@ -1,6 +1,6 @@
 import { drizzle } from "drizzle-orm/neon-http";
-import * as schema from "./schema";
+import * as schema from "./schema/index";
 
 export const db = drizzle(process.env.DATABASE_URL!, {
-  schema: schema,
+  schema,
 });
